refactor(alias): update to single page editation
