+++ src/panel/index.ts
@@ -115,8 +115,8 @@ const main = async () => {
         { path: '/stats/bits', name: 'BitsStats', component: () => import('./views/stats/bits.vue') },
 
         { path: '/manage/alias/', redirect: '/manage/alias/list' },
-        { path: '/manage/alias/list', name: 'aliasManagerList', component: () => import('./views/managers/alias/alias-list.vue') },
-        { path: '/manage/alias/edit/:id?', name: 'aliasManagerEdit', component: () => import('./views/managers/alias/alias-edit.vue') },
+        { path: '/manage/alias/list', name: 'aliasManager', component: () => import('./views/managers/alias.vue') },
+        { path: '/manage/alias/edit/:id?', name: 'aliasManagerEdit', component: () => import('./views/managers/alias.vue') },
         { path: '/manage/commands/', redirect: '/manage/commands/list' },
         { path: '/manage/commands/list', name: 'CommandsManagerList', component: () => import('./views/managers/commands/commands-list.vue') },
         { path: '/manage/commands/edit/:id?', name: 'CommandsManagerEdit', component: () => import('./views/managers/commands/commands-edit.vue') },
